requirements.txt
accelerate==0.30.1
antlr4-python3-runtime==4.9.3
av==12.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
diffusers==0.27.2
einops==0.8.0
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.5.0
fvcore==0.1.5.post20221221
huggingface-hub==0.23.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
iopath==0.1.9
Jinja2==3.1.4
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
# networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pip==24.0
portalocker==2.3.0
psutil==5.9.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
safetensors==0.4.3
# scikit-image==0.23.2
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
sympy==1.12
tabulate==0.9.0
termcolor==2.1.0
tifffile==2024.5.10
timm==1.0.3
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.2
transformers==4.40.2
trimesh==4.3.2
typing_extensions==4.11.0
urllib3==2.2.1
wheel==0.43.0
yacs==0.1.6
zipp==3.18.1
